requirements.txt
fastapi
fastapi[standard]
pyjwt
passlib
bcrypt
python-dotenv
sqlmodel
pydantic-settings
alembic
pyodbc  
aioodbc
boto3
awscli-local[ver1]
pandas
gradio_client
langchain
langchain_community
llama-cpp-python
pypdf
sentence-transformers
xlsxwriter


